Fix: Fixes bug that could keep a refresh rate in persistence even if the UseAutoRefreshRate property was false

+++ Program.cs
@@ -64,10 +64,10 @@ namespace AutoRefreshHDR
 
                         if (hdrActivated || refreshRateChange)
                         {
-                            if (hdrActivated)
+                            if (displayConfig.UseAutoHDR && hdrActivated)
                                 HDRSwitchOff();
 
-                            if (refreshRateChange)
+                            if (displayConfig.UseAutoRefreshRate && refreshRateChange)
                             {
                                 ChangeRefreshRate(currentRefreshRate);
                                 DeleteRefreshRatePersisted();
